Display date and time of earthquake  On branch L1_003_Display_date_and_time_of_earthquake  Changes to be committed: 	modified:   .idea/misc.xml 	modified:   app/src/main/java/com/oz_heng/apps/android/quakereport/Earthquake.java 	modified:   app/src/main/java/com/oz_heng/apps/android/quakereport/EarthquakeActivity.java 	modified:   app/src/main/java/com/oz_heng/apps/android/quakereport/EarthquakeAdapter.java 	modified:   app/src/main/java/com/oz_heng/apps/android/quakereport/Helper.java 	modified:   app/src/main/java/com/oz_heng/apps/android/quakereport/QueryUtils.java 	modified:   app/src/main/res/layout/earthquake_list_item.xml

+++ app/src/main/java/com/oz_heng/apps/android/quakereport/QueryUtils.java
@@ -9,8 +9,6 @@ import org.json.JSONObject;
 import java.util.ArrayList;
 import java.util.Date;
 
-import static com.oz_heng.apps.android.quakereport.Helper.stringToDate;
-
 /**
  * Helper methods related to requesting and receiving earthquake data from USGS.
  */
@@ -64,7 +62,7 @@ public final class QueryUtils {
 
                 double magnitude = properties.getDouble("mag");
                 String place = properties.getString("place");
-                Date date = new Date( properties.getInt("time") );
+                Date date = new Date( properties.getLong("time") );
 
                 earthquakes.add( new Earthquake(magnitude, place, date) );
             }
